Fix a serious bug in the filters for getWallPosts.
There was a problem with the get wall post retrieving posts it should have filtered. this was due to a problem with the conditions array. author id would cause the user id to be false if not provided. Added to new filters aswell. buid or blocked user id, and baid or blocked user id.

Signed-off-by: Robert Hurst <Robert@ThinkTankDesign.ca>